[GTK] Fix some test failures in ATK selection handling. https://bugs.webkit.org/show_bug.cgi?id=168369 <rdar://problem/30534881>
Reviewed by Joanmarie Diggs.

Source/WebCore:

In r208479, selectionBelongsToObject was changed to return false if the
intersectsNode call returns an exception.

In particular, this caused accessibility/gtk/text-at-offset-textarea.html
to fail. In this test, the selection is situated in the shadow DOM of the
textarea, while the node that is checked for intersection is the textarea
itself. In line with the standard, intersectsNode returns an exception in
this case.

This caused webkitAccessibleText{Word, Line, Sentence}ForBoundary to stop
returning the expected text in the tested case. Removing this check fixes
the test, along with some others.

Tests: accessibility/gtk/text-at-offset-textarea.html
       accessibility/gtk/text-at-offset-textinput.html
       accessibility/selected-text-range-aria-elements.html
       accessibility/textarea-selected-text-range.html

* accessibility/atk/WebKitAccessibleInterfaceText.cpp:
(getSelectionOffsetsForObject): Remove the selectionBelongsToObject() call.

LayoutTests:

Update test expectations for tests that pass again.

* platform/gtk/TestExpectations:


Identifier: 197875@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@227523 268f45cc-cd09-0410-ab3c-d52691b4dbfc