REGRESSiON (r226492): Crash under Element::absoluteEventBounds() on a SVGPathElement which has not been laid out yet https://bugs.webkit.org/show_bug.cgi?id=182185 rdar://problem/36836262
Reviewed by Zalan Bujtas.

Document::absoluteRegionForEventTargets() can fire when layout is dirty, and SVGPathElement's path() can be null if it
hasn't been laid out yet. So protect against a null path in getBBox().

Not easily testable because internals.nonFastScrollableRects() forces layout, and the crash depends on the timing of
absoluteRegionForEventTargets().

* svg/SVGPathElement.cpp:
(WebCore::SVGPathElement::getBBox):

Identifier: 198009@main
git-svn-id: https://svn.webkit.org/repository/webkit/trunk@227697 268f45cc-cd09-0410-ab3c-d52691b4dbfc